refactor: 참여자 주문 삭제 시 본인이 주문한게 맞는지 로직 추가 (#172)

+++ src/main/java/com/example/surveyapp/domain/order/service/OrderService.java
@@ -115,6 +115,10 @@ public class OrderService {
         Order order = orderRepository.findByIdAndIsDeletedFalse(id)
                 .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_ORDER));
 
+        if (!order.getUser().getId().equals(userId)) {
+            throw new CustomException(ErrorCode.NOT_SAME_ORDER_USER);
+        }
+
         order.delete();
 
     }
